history fucking works!!!

+++ KeePark/Controllers/ReserveSpotsController.cs
@@ -11,26 +11,83 @@ using System.Net;
 using System.IO;
 using Newtonsoft.Json;
 using System.Text;
+using System.Security.Claims;
+using Microsoft.AspNetCore.Identity;
 
 namespace KeePark.Controllers
 {
     public class ReserveSpotsController : Controller
     {
+        private readonly UserManager<GeneralUser> _userManager;
         private readonly KeeParkContext _context;
         private readonly IdentityContext _identitycontext;
 
-        public ReserveSpotsController(KeeParkContext context, IdentityContext identitycontext)
+        public ReserveSpotsController(KeeParkContext context, IdentityContext identitycontext, UserManager<GeneralUser> userManager)
         {
             _context = context;
             _identitycontext = identitycontext;
+            _userManager = userManager;
         }
 
-        // GET: ReserveSpots
-        public async Task<IActionResult> Index()
+
+        public ActionResult Orders(string datepicker, string employee, string visitor)
         {
-            var keeParkContext = _context.ReserveSpot.Include(r => r.Spot).Include(r => r.User);
-            return View(await keeParkContext.ToListAsync());
+            var userid = _userManager.GetUserId(HttpContext.User);
+            GeneralUser user = _userManager.FindByIdAsync(userid).Result;
+            var res = (from reservations in _context.ReserveSpot
+                       where reservations.UserID == user.UID
+                       select reservations).Include(r => r.Spot);
+
+
+
+            //filter data to get an filtered list
+
+            return PartialView("_ResultTable", res);
         }
+
+
+        // GET: ReserveSpots
+        public IActionResult Index(string spotsName, DateTime resDate, string carNumber)
+        {
+
+
+            var userid = _userManager.GetUserId(HttpContext.User);
+            GeneralUser user = _userManager.FindByIdAsync(userid).Result;
+            var res = (from reservations in _context.ReserveSpot
+                       where reservations.UserID == user.UID
+                       select reservations).Include(r => r.Spot);
+
+            var reserves = from r in res select r;
+
+            // Smart Search
+            if (!String.IsNullOrEmpty(spotsName))
+            {
+                reserves = reserves.Where(a => a.Spot.SpotName.Contains(spotsName));
+            }
+
+            if (resDate != DateTime.MinValue)
+            {
+                reserves = reserves.Where(a => a.ReservationDate == resDate);
+            }
+
+            if (!String.IsNullOrEmpty(carNumber))
+            {
+                reserves = reserves.Where(a => a.carNumber.Contains(carNumber));
+            }
+
+            return View(reserves.ToList());
+        }
+
+        public IActionResult FutureOrders()
+        {
+            var userid = _userManager.GetUserId(HttpContext.User);
+            GeneralUser user = _userManager.FindByIdAsync(userid).Result;
+            var reserve = (from reservations in _context.ReserveSpot
+                           where reservations.UserID == user.UID && reservations.ReservationDate > DateTime.Now
+                           select reservations).Include(r => r.Spot);
+            return View(reserve);
+        }
+
 
         // GET: ReserveSpots/Details/5
         public async Task<IActionResult> Details(Guid? id)
@@ -42,7 +99,6 @@ namespace KeePark.Controllers
 
             var reserveSpot = await _context.ReserveSpot
                 .Include(r => r.Spot)
-                .Include(r => r.User)
                 .FirstOrDefaultAsync(m => m.ReserveSpotID == id);
             if (reserveSpot == null)
             {
@@ -55,16 +111,17 @@ namespace KeePark.Controllers
         //  [Route("ReserveSpots/Create/{parkingSpotID}")]
         public IActionResult Create([FromRoute]int parkingSpotID)
         {
+            //getting car number for the view
             var carN = (from mycar in _identitycontext.GeneralUser
                         where mycar.Email == User.Identity.Name
                         select mycar.CarNumber).FirstOrDefault();
+            //getting user id
             var userid = (from myid in _identitycontext.GeneralUser
                           where myid.Email == User.Identity.Name
                           select myid.UID).FirstOrDefault();
 
             var viewModel = new ReserveSpot
             {
-
                 carNumber = carN,
                 ReservationDate = System.DateTime.Now,
                 CreatedOn = System.DateTime.Now,
@@ -86,116 +143,89 @@ namespace KeePark.Controllers
             if (ModelState.IsValid)
             {
 
-                if ((!(reserveSpot.ReservationDate < System.DateTime.Now)) || (reserveSpot.ReservationDate.Date == DateTime.Today))
+                //checks if the reservation date didn't pass or if the the reservation date is today
+                if ((reserveSpot.ReservationDate.Date >= DateTime.Today))
                 {
-
+                    //checks if the reservation hour is valid for today's date
                     if ((reserveSpot.ReservationHour <= System.DateTime.Now.Hour) && ((reserveSpot.ReservationDate.Date == DateTime.Today)))
                         return RedirectToAction(nameof(InvalidHour));
 
 
+                    //getting the id of current user
+                    var currentUser = (from userID in _identitycontext.GeneralUser
+                                       where userID.Id == User.FindFirstValue(ClaimTypes.NameIdentifier)
+                                       select userID.UID).FirstOrDefault();
 
 
-                    var user = (from bla in _identitycontext.GeneralUser
-                                where bla.UserName == User.Identity.Name
-                                select bla).FirstOrDefault();
-                    var userid = (from bla in _identitycontext.GeneralUser
-                                  where bla.UserName == User.Identity.Name
-                                  select bla.UID).FirstOrDefault();
+
 
                     reserveSpot.ReserveSpotID = Guid.NewGuid();
                     reserveSpot.CreatedOn = System.DateTime.Now;
                     reserveSpot.Spot = _context.ParkingSpot.FirstOrDefault(u => u.ParkingSpotID == parkingSpotID);
-
-                    int count = 0;
-                    foreach (var t in _context.ReserveSpot)
-                    {
-                        if (userid == t.UserID)
-                        {
-                            count++;
-                        }
-
-                    }
-
-                    if (count == 0 && !(_context.GeneralUser.Contains(user)))
-                    {
-                        reserveSpot.User = _identitycontext.GeneralUser.FirstOrDefault(u => u.UserName == User.Identity.Name);
-                        reserveSpot.UserID = reserveSpot.User.UID;
-
-                    }
-                    else
-                    {
-
-                        reserveSpot.UserID = userid;
-                    }
+                    reserveSpot.UserID = currentUser;
 
 
+
+                    //get all the reservations for the wanted day for a specific spot spot
                     var myreservations = (from reservations in _context.ReserveSpot
-                                          where reservations.Spot.ParkingSpotID == parkingSpotID  //replace
+                                          where reservations.Spot.ParkingSpotID == parkingSpotID && reservations.ReservationDate == reserveSpot.ReservationDate  //replace
                                           select reservations);
-                    var m = (from t in myreservations
-                             where t.ReservationDate == reserveSpot.ReservationDate
-                             select t);
 
-                    foreach (var t in m)
-                    {
-                        if ((t.ReservationHour <= reserveSpot.ReservationHour) && (reserveSpot.ReservationHour < t.ReservationHour + t.Duration))
-                        {
-                            return RedirectToAction(nameof(SorrySpotIsTaken));
-                        }
-
-                    }
-
-                    var reserves = from res in _context.ReserveSpot
-                                   where res.ReservationDate == reserveSpot.ReservationDate && res.SpotID == parkingSpotID
-                                   select res;
-                    foreach (var r in reserves)
-                    {
-
-                        if ((reserveSpot.ReservationHour < r.ReservationHour + r.Duration && reserveSpot.ReservationHour >= r.ReservationHour))
-                        {
-                            return RedirectToAction(nameof(SorrySpotIsTaken));
-                        }
-                        if ((reserveSpot.ReservationHour + reserveSpot.Duration <= r.ReservationHour + r.Duration) && (reserveSpot.ReservationHour + reserveSpot.Duration > r.ReservationHour))
-                        {
-                            return RedirectToAction(nameof(SorrySpotIsTaken));
-                        }
-                        if ((reserveSpot.ReservationHour <= r.ReservationHour) && (reserveSpot.ReservationHour + reserveSpot.Duration >= r.ReservationHour + r.Duration))
-                        {
-                            return RedirectToAction(nameof(SorrySpotIsTaken));
-                        }
-
-                    }
+                    //validation that the reservation made for the desired date
                     if (reserveSpot.ReservationHour + reserveSpot.Duration > 24)
                     {
                         return RedirectToAction(nameof(SorryChooseDurationForTheDayYouSelected));
                     }
 
 
+                    foreach (var r in myreservations)
+                    {
+                        //case1: check if the desired res hour is between an already reserved time range
+                        if ((reserveSpot.ReservationHour < r.ReservationHour + r.Duration && reserveSpot.ReservationHour >= r.ReservationHour))
+                        {
+                            return RedirectToAction(nameof(SorrySpotIsTaken));
+                        }
+
+                        //case2: the desired res hour is an availble hour but the duration slip into other reservation and its end time is  before the already existing reservation fullfiled
+                        if ((reserveSpot.ReservationHour + reserveSpot.Duration <= r.ReservationHour + r.Duration) && (reserveSpot.ReservationHour + reserveSpot.Duration > r.ReservationHour))
+                        {
+                            return RedirectToAction(nameof(SorrySpotIsTaken));
+                        }
+                        //case3: the desired res hour is an availble hour but the duration slip into other reservation
+                        if ((reserveSpot.ReservationHour <= r.ReservationHour) && (reserveSpot.ReservationHour + reserveSpot.Duration >= r.ReservationHour + r.Duration))
+                        {
+                            return RedirectToAction(nameof(SorrySpotIsTaken));
+                        }
+
+                    }
 
 
                     reserveSpot.Spot.NunOfOrders++;
                     _context.Add(reserveSpot);
-
-
                     await _context.SaveChangesAsync();
 
-                    GeneralUser currentUser = new GeneralUser();
-                    currentUser = (from cuser in _identitycontext.GeneralUser
-                                   where cuser.UID == reserveSpot.UserID
-                                   select cuser).FirstOrDefault();
+                    //GeneralUser thisUser = new GeneralUser();
+                    var username = User.Identity.Name;
+                    var user = await _identitycontext.GeneralUser.FirstOrDefaultAsync(u => u.UserName.Equals(username));
+                    //thisUser = (from cuser in _identitycontext.GeneralUser
+                    //              where cuser.Id == reserveSpot.UserID
+                    //             select cuser).FirstOrDefault();
+                    if (user == null)
+                        return NotFound();
 
                     // this section is to save the history of spots reserved to serve the ML algo
-                    string uHistory = currentUser.History;
-                    if (!string.IsNullOrEmpty(uHistory)) {
+                    string uHistory = user.History;
+                    if (!string.IsNullOrEmpty(uHistory))
+                    {
                         uHistory += ",";
                     }
                     uHistory += reserveSpot.SpotID.ToString();
-                    currentUser.History = uHistory;
+                    user.History = uHistory;
 
                     try
                     {
-                        _identitycontext.Update(currentUser);
-                        await _context.SaveChangesAsync();
+                        _identitycontext.Update(user);
+                        await _identitycontext.SaveChangesAsync();
                     }
                     catch (DbUpdateConcurrencyException)
                     {
@@ -213,10 +243,10 @@ namespace KeePark.Controllers
                 }
 
 
-            }
 
-            // ViewData["UserID"] = new SelectList(_identitycontext.GeneralUser, "UID", "UID", reserveSpot.UserID);
+            }
             return View(reserveSpot);
+
         }
 
 
@@ -272,59 +302,63 @@ namespace KeePark.Controllers
 
             if (ModelState.IsValid)
             {
-                if ((!(reserveSpot.ReservationDate < System.DateTime.Now)) || (reserveSpot.ReservationDate.Date == DateTime.Today))
+                if ((reserveSpot.ReservationDate.Date >= DateTime.Today))
                 {
+                    //checks if the reservation hour is valid for today's date
                     if ((reserveSpot.ReservationHour <= System.DateTime.Now.Hour) && ((reserveSpot.ReservationDate.Date == DateTime.Today)))
                         return RedirectToAction(nameof(InvalidHour));
 
                     try
                     {
+                        var userid = (from users in _identitycontext.GeneralUser
+                                      where users.UserName == User.Identity.Name
+                                      select users.UID).FirstOrDefault();
                         var spotid = (from myid in _context.ReserveSpot
                                       where myid.ReserveSpotID == id
                                       select myid.SpotID).FirstOrDefault();
-                        reserveSpot.User = _identitycontext.GeneralUser.FirstOrDefault(u => u.UserName == User.Identity.Name);
-                        reserveSpot.UserID = reserveSpot.User.UID;
+
+                        reserveSpot.UserID = userid;
                         reserveSpot.Spot = _context.ParkingSpot.FirstOrDefault(u => u.ParkingSpotID == spotid);
                         reserveSpot.SpotID = reserveSpot.Spot.ParkingSpotID;
                         reserveSpot.CreatedOn = System.DateTime.Now;
 
+
                         var res = from reserve in _context.ReserveSpot
                                   where reserve.ReserveSpotID != id
                                   select reserve;
+
                         var spots = from sp in res
                                     where sp.Spot.ParkingSpotID == reserveSpot.SpotID && sp.ReservationDate == reserveSpot.ReservationDate
                                     select sp;
 
-                        foreach (var t in spots)
+                        //validation that the reservation made for the desired date
+                        if (reserveSpot.ReservationHour + reserveSpot.Duration > 24)
                         {
-                            if ((t.ReservationHour <= reserveSpot.ReservationHour) && (reserveSpot.ReservationHour < t.ReservationHour + t.Duration))
-                            {
-                                return RedirectToAction(nameof(SorrySpotIsTaken));
-                            }
+                            return RedirectToAction(nameof(SorryChooseDurationForTheDayYouSelected));
                         }
-
 
                         foreach (var r in spots)
                         {
 
+                            //case1: check if the desired res hour is between an already reserved time range
                             if ((reserveSpot.ReservationHour < r.ReservationHour + r.Duration && reserveSpot.ReservationHour >= r.ReservationHour))
                             {
                                 return RedirectToAction(nameof(SorrySpotIsTaken));
                             }
+
+                            //case2: the desired res hour is an availble hour but the duration slip into other reservation and its end time is  before the already existing reservation fullfiled
                             if ((reserveSpot.ReservationHour + reserveSpot.Duration <= r.ReservationHour + r.Duration) && (reserveSpot.ReservationHour + reserveSpot.Duration > r.ReservationHour))
                             {
                                 return RedirectToAction(nameof(SorrySpotIsTaken));
                             }
+                            //case3: the desired res hour is an availble hour but the duration slip into other reservation
                             if ((reserveSpot.ReservationHour <= r.ReservationHour) && (reserveSpot.ReservationHour + reserveSpot.Duration >= r.ReservationHour + r.Duration))
                             {
                                 return RedirectToAction(nameof(SorrySpotIsTaken));
                             }
 
                         }
-                        if (reserveSpot.ReservationHour + reserveSpot.Duration > 24)
-                        {
-                            return RedirectToAction(nameof(SorryChooseDurationForTheDayYouSelected));
-                        }
+
 
                         _context.Update(reserveSpot);
                         await _context.SaveChangesAsync();
@@ -369,7 +403,6 @@ namespace KeePark.Controllers
 
             var reserveSpot = await _context.ReserveSpot
                 .Include(r => r.Spot)
-                .Include(r => r.User)
                 .FirstOrDefaultAsync(m => m.ReserveSpotID == id);
             if (reserveSpot == null)
             {
@@ -385,7 +418,18 @@ namespace KeePark.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> DeleteConfirmed(Guid id)
         {
+
             var reserveSpot = await _context.ReserveSpot.FindAsync(id);
+            //cannot delete a fullfiled res
+            if ((reserveSpot.ReservationDate.Date < DateTime.Today) || ((reserveSpot.ReservationDate.Date == DateTime.Today) && (reserveSpot.ReservationHour <= System.DateTime.Now.Hour)))
+            {
+                return RedirectToAction(nameof(CanNotDelete));
+            }
+
+            var spot = (from res in _context.ReserveSpot
+                        where res.ReserveSpotID == id
+                        select res.Spot).FirstOrDefault();
+            spot.NunOfOrders--;
             _context.ReserveSpot.Remove(reserveSpot);
             await _context.SaveChangesAsync();
             return RedirectToAction(nameof(Index));
@@ -401,7 +445,10 @@ namespace KeePark.Controllers
         {
             return View();
         }
-
+        public IActionResult CanNotDelete()
+        {
+            return View();
+        }
